added try catch for negative amount

+++ PaymentGateway/PaymentGateway/IPaymentGateway.cs
@@ -6,6 +6,6 @@ namespace PaymentGateway
 {
     public interface IPaymentGateway
     {
-        PaymentResult Pay(double amount);
+        PaymentResult Pay(decimal amount);
     }
 }
